+++ src/RcsbFvWeb/RcsbFvFactories/RcsbFvBlockFactory/AlignmentBlockFactory.ts
@@ -7,14 +7,14 @@ import {BlockFactoryInterface} from "./BlockFactoryInterface";
 import {RcsbFvRowConfigInterface} from "@rcsb/rcsb-saguaro";
 import {TrackFactoryInterface} from "../RcsbFvTrackFactory/TrackFactoryInterface";
 
-export class AlignmentBlockFactory implements BlockFactoryInterface<[AlignmentRequestContextType, AlignmentResponse],[AlignmentRequestContextType, TargetAlignment, AlignmentResponse]> {
+export class AlignmentBlockFactory implements BlockFactoryInterface<[AlignmentRequestContextType, AlignmentResponse],[AlignmentRequestContextType, TargetAlignment, AlignmentResponse, number]> {
 
     readonly trackFactory: TrackFactoryInterface<[AlignmentRequestContextType, TargetAlignment]>;
-    readonly trackConfigModifier?: (alignmentContext: AlignmentRequestContextType, targetAlignment: TargetAlignment, alignmentResponse: AlignmentResponse) => Promise<Partial<RcsbFvRowConfigInterface>>;
+    readonly trackConfigModifier?: (alignmentContext: AlignmentRequestContextType, targetAlignment: TargetAlignment, alignmentResponse: AlignmentResponse, indexResponse: number) => Promise<Partial<RcsbFvRowConfigInterface>>;
 
     constructor(
         alignmentTrackFactory: TrackFactoryInterface<[AlignmentRequestContextType, TargetAlignment]>,
-        trackModifier?: (alignmentContext: AlignmentRequestContextType, targetAlignment: TargetAlignment, alignmentResponse: AlignmentResponse) => Promise<Partial<RcsbFvRowConfigInterface>>
+        trackModifier?: (alignmentContext: AlignmentRequestContextType, targetAlignment: TargetAlignment, alignmentResponse: AlignmentResponse, indexResponse: number) => Promise<Partial<RcsbFvRowConfigInterface>>
     ) {
         this.trackFactory = alignmentTrackFactory;
         this.trackConfigModifier = trackModifier;
@@ -24,14 +24,14 @@ export class AlignmentBlockFactory implements BlockFactoryInterface<[AlignmentRe
         alignmentRequestContext = {...alignmentRequestContext, querySequence: alignmentData.query_sequence ?? undefined};
         if(!alignmentData.target_alignment)
             return [];
-        return (await Promise.all<undefined|RcsbFvRowConfigInterface>(alignmentData.target_alignment.map(async alignment=>{
+        return (await Promise.all<undefined|RcsbFvRowConfigInterface>(alignmentData.target_alignment.map(async (alignment, index)=>{
             if (alignmentRequestContext.filterByTargetContains != null && !alignment?.target_id?.includes(alignmentRequestContext.filterByTargetContains))
                 return;
             if (alignment?.target_sequence == null)
                 return;
            return {
                ... await this.trackFactory.getTrack(alignmentRequestContext, alignment),
-               ... await this.trackConfigModifier?.(alignmentRequestContext,alignment, alignmentData)
+               ... await this.trackConfigModifier?.(alignmentRequestContext,alignment, alignmentData, index)
            };
         }))).filter((x): x is RcsbFvRowConfigInterface=>x!=undefined);
     }